refactor : api 명 변경 캐싱을 고려한 api 설계
